+++ compiler/rustc_query_impl/src/values.rs
@@ -8,10 +8,10 @@ use rustc_errors::{Applicability, MultiSpan, pluralize, struct_span_code_err};
 use rustc_hir as hir;
 use rustc_hir::def::{DefKind, Res};
 use rustc_middle::dep_graph::dep_kinds;
+use rustc_middle::query::CycleError;
 use rustc_middle::query::plumbing::CyclePlaceholder;
 use rustc_middle::ty::{self, Representability, Ty, TyCtxt};
 use rustc_middle::{bug, span_bug};
-use rustc_query_system::query::CycleError;
 use rustc_span::def_id::LocalDefId;
 use rustc_span::{ErrorGuaranteed, Span};
